fix: correctly calcualte outline max width
From experimenting with OneNote it seems like the max-width
of an outline is calculated differently if the layout size
is *not* customized by the user. In this case, a minimum width
is used that seems to be device dependent. In our case we're
generating a HTML file so 13 inches seems like a good choice.