fix: correctly render paragraph indentations
Previously we assumed that all indentations are consecutive, i.e.
level 1 is always followed by level 2. However, this is not always
the case as the `child_level` property may indicate that child
nodes should be indented by more than one level.

In addition, we didn't use the `Outline::indents()` to calcualte
the indentation width, instead assuming that all indents have
a width of 1.5 inch.